requirements.txt
requests==2.32.3
google-auth==2.27.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
